#12 - prepare for cqfn
