+++ LICENSE.txt
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2020 g4s8
+Copyright (c) 2020 cqfn.org
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files
