--- src/main/java/wtf/g4s8/rio/file/ShutdownOnExit.java
@@ -1,63 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 g4s8
- *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files
- * (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights * to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software,
- * and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
- * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- * OTHER DEALINGS IN THE SOFTWARE.
- */
-package wtf.g4s8.rio.file;
-
-import java.util.concurrent.ExecutorService;
-
-/**
- * Runnable decorator which shutdowns {@link ExecutorService} on exit.
- * @since 0.1
- */
-final class ShutdownOnExit implements Runnable {
-
-    /**
-     * Executor service.
-     */
-    private final ExecutorService exec;
-
-    /**
-     * Origin runnable.
-     */
-    private final Runnable origin;
-
-    /**
-     * Wrap origin runnable.
-     * @param origin Origin runnable
-     * @param exec Executor to shutdown
-     */
-    ShutdownOnExit(final Runnable origin, final ExecutorService exec) {
-        this.exec = exec;
-        this.origin = origin;
-    }
-
-    @Override
-    public void run() {
-        try {
-            this.origin.run();
-        } finally {
-            this.exec.shutdown();
-        }
-    }
-}
